# Introduces more standard types wasm.Index, wasm.NameMap etc.