--- a/wasm/module.go
+++ b/wasm/module.go
@@ -31,39 +31,43 @@ type Module struct {
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#import-section%E2%91%A0
 	ImportSection []*Import
+
 	// FunctionSection contains the index in TypeSection of each function defined in this module.
 	//
-	// Function indexes are offset by any imported functions because the function index space begins with imports,
-	// followed by ones defined in this module. Moreover, the FunctionSection is index correlated with the CodeSection.
+	// Note: The function Index namespace begins with imported functions and ends with those defined in this module.
+	// For example, if there are two imported functions and one defined in this module, the function Index 3 is defined
+	// in this module at FunctionSection[0].
 	//
-	// For example, if there are two imported functions and three defined in this module, we expect the CodeSection to
-	// have a length of three and the function at index 3 to be defined in this module. Its type would be at
-	// TypeSection[FunctionSection[0]], while its locals and body are at CodeSection[0].
+	// Note: FunctionSection is index correlated with the CodeSection. If given the same position, ex. 2, a function
+	// type is at TypeSection[FunctionSection[2]], while its locals and body are at CodeSection[2].
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#function-section%E2%91%A0
-	FunctionSection []uint32
+	FunctionSection []Index
+
 	// TableSection contains each table defined in this module.
 	//
-	// Table indexes are offset by any imported tables because the table index space begins with imports, followed by
-	// ones defined in this module. For example, if there are two imported tables and one defined in this module, the
-	// table at index 3 is defined in this module at TableSection[0].
+	// Note: The table Index namespace begins with imported tables and ends with those defined in this module.
+	// For example, if there are two imported tables and one defined in this module, the table Index 3 is defined in
+	// this module at TableSection[0].
 	//
 	// Note: Version 1.0 (MVP) of the WebAssembly spec allows at most one table definition per module, so the length of
 	// the TableSection can be zero or one.
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#table-section%E2%91%A0
 	TableSection []*TableType
+
 	// MemorySection contains each memory defined in this module.
 	//
-	// Memory indexes are offset by any imported memories because the memory index space begins with imports, followed
-	// by ones defined in this module. For example, if there are two imported memories and one defined in this module,
-	// the memory at index 3 is defined in this module at MemorySection[0].
+	// Note: The memory Index namespace begins with imported memories and ends with those defined in this module.
+	// For example, if there are two imported memories and one defined in this module, the memory Index 3 is defined in
+	// this module at TableSection[0].
 	//
 	// Note: Version 1.0 (MVP) of the WebAssembly spec allows at most one memory definition per module, so the length of
 	// the MemorySection can be zero or one.
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#memory-section%E2%91%A0
 	MemorySection []*MemoryType
+
 	// GlobalSection contains each global defined in this module.
 	//
 	// Global indexes are offset by any imported globals because the global index space begins with imports, followed by
@@ -72,21 +76,26 @@ type Module struct {
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#global-section%E2%91%A0
 	GlobalSection []*Global
+
 	ExportSection map[string]*Export
+
 	// StartSection is the index of a function to call before returning from Store.Instantiate.
 	//
-	// Note: The function index space begins with any ImportKindFunc in ImportSection, then the FunctionSection.
-	// For example, if there are two imported functions and three defined in this module, the index space is five.
-	// Note: This is a pointer to avoid conflating no start section with the valid index zero.
+	// Note: The index here is not the position in the FunctionSection, rather in the function index namespace, which
+	// begins with imported functions.
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#start-section%E2%91%A0
-	StartSection   *uint32
+	StartSection *Index
+
 	ElementSection []*ElementSegment
+
 	// CodeSection is index-correlated with FunctionSection and contains each function's locals and body.
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#code-section%E2%91%A0
 	CodeSection []*Code
+
 	DataSection []*DataSegment
+
 	// NameSection is set when the custom section "name" was successfully decoded from the binary format.
 	//
 	// Note: This is the only custom section defined in the WebAssembly 1.0 (MVP) Binary Format. Others are in
@@ -94,6 +103,7 @@ type Module struct {
 	//
 	// See https://www.w3.org/TR/wasm-core-1/#name-section%E2%91%A0
 	NameSection *NameSection
+
 	// CustomSections is set when at least one non-standard, or otherwise unsupported custom section was found in the
 	// binary format.
 	//
@@ -102,6 +112,15 @@ type Module struct {
 	// See https://www.w3.org/TR/wasm-core-1/#custom-section%E2%91%A0
 	CustomSections map[string][]byte
 }
+
+// Index is the offset in an index namespace, not necessarily an absolute position in a Module section. This is because
+// index namespaces are often preceded by a corresponding type in the Module.ImportSection.
+//
+// For example, the function index namespace starts with any ImportKindFunc in the Module.TypeSection followed by the
+// Module.FunctionSection
+//
+// See https://www.w3.org/TR/wasm-core-1/#binary-index
+type Index = uint32
 
 type FunctionType struct {
 	Params, Results []ValueType
@@ -171,7 +190,7 @@ type Import struct {
 	// Module is the possibly empty secondary namespace of this import
 	Name string
 	// DescFunc is the index in Module.TypeSection when Kind equals ImportKindFunc
-	DescFunc uint32
+	DescFunc Index
 	// DescTable is the inlined TableType when Kind equals ImportKindTable
 	DescTable *TableType
 	// DescMem is the inlined MemoryType when Kind equals ImportKindMemory
@@ -225,11 +244,11 @@ type Export struct {
 	Name string
 	// Index is the index of the definition to export, the index namespace is by Kind
 	// Ex. If ExportKindFunc, this is an index to ModuleInstance.Functions
-	Index uint32
+	Index Index // TODO: can you export an import? If so, rewrite the doc ^^ to say "index namespace"
 }
 
 type ElementSegment struct {
-	TableIndex uint32
+	TableIndex Index
 	OffsetExpr *ConstantExpression
 	Init       []uint32
 }
@@ -241,40 +260,68 @@ type Code struct {
 }
 
 type DataSegment struct {
-	MemoryIndex      uint32 // supposed to be zero
+	MemoryIndex      Index // supposed to be zero
 	OffsetExpression *ConstantExpression
 	Init             []byte
 }
 
 // NameSection represent the known custom name subsections defined in the WebAssembly Binary Format
+//
+// Note: This can be nil if no names were decoded for any reason including configuration.
 // See https://www.w3.org/TR/wasm-core-1/#name-section%E2%91%A0
-// See https://github.com/tetratelabs/wazero/issues/134 about adding this to Module
 type NameSection struct {
 	// ModuleName is the symbolic identifier for a module. Ex. math
-	// The corresponding subsection is subsectionIDModuleName.
+	//
+	// Note: This can be empty for any reason including configuration.
 	ModuleName string
+
 	// FunctionNames is an association of a function index to its symbolic identifier. Ex. add
-	// The corresponding subsection is subsectionIDFunctionNames.
 	//
 	// * the key (idx) is in the function namespace, where module defined functions are preceded by imported ones.
+	// See https://www.w3.org/TR/wasm-core-1/#functions%E2%91%A7
 	//
 	// Ex. Assuming the below text format is the second import, you would expect FunctionNames[1] = "mul"
 	//	(import "Math" "Mul" (func $mul (param $x f32) (param $y f32) (result f32)))
 	//
-	// Note: FunctionNames are a map because the specification requires function indices to be unique. These are sorted
-	// during EncodeData
-	FunctionNames map[uint32]string
+	// Note: FunctionNames are only used for debugging. At runtime, functions are called based on raw numeric index.
+	// Note: This can be nil for any reason including configuration.
+	FunctionNames NameMap
 
-	// LocalNames is an association of a function index to any locals which have a symbolic identifier. Ex. add x
-	// The corresponding subsection is subsectionIDLocalNames.
+	// LocalNames contains symbolic names for function parameters or locals that have one.
 	//
-	// * the key (funcIndex) is in the function namespace, where module defined functions are preceded by imported ones.
-	// * the second key (idx) is in the local namespace, where parameters precede any function locals.
+	// Note: In the Text Format, function local names can inherit parameter names from their type. Ex.
+	//  * (module (import (func (param $x i32) (param i32))) (func (type 0))) = [{0, {x,0}}]
+	//  * (module (import (func (param i32) (param $y i32))) (func (type 0) (local $z i32))) = [0, [{y,1},{z,2}]]
+	//  * (module (func (param $x i32) (local $y i32) (local $z i32))) = [{x,0},{y,1},{z,2}]
 	//
-	// Ex. Assuming the below text format is the second import, you would expect LocalNames[1][1] = "y"
-	//	(import "Math" "Mul" (func $mul (param $x f32) (param $y f32) (result f32)))
-	//
-	// Note: LocalNames are a map because the specification requires both function and local indices to be unique. These
-	// are sorted during EncodeData
-	LocalNames map[uint32]map[uint32]string
+	// Note: LocalNames are only used for debugging. At runtime, locals are called based on raw numeric index.
+	// Note: This can be nil for any reason including configuration.
+	LocalNames IndirectNameMap
+}
+
+// NameMap associates an index with any associated names.
+//
+// Note: Often the index namespace bridges multiple sections. For example, the function index namespace starts with any
+// ImportKindFunc in the Module.TypeSection followed by the Module.FunctionSection
+//
+// Note: NameMap is unique by NameAssoc.Index, but NameAssoc.Name needn't be unique.
+// Note: When encoding in the Binary format, this must be ordered by NameAssoc.Index
+// See https://www.w3.org/TR/wasm-core-1/#binary-namemap
+type NameMap []*NameAssoc
+
+type NameAssoc struct {
+	Index Index
+	Name  string
+}
+
+// IndirectNameMap associates an index with an association of names.
+//
+// Note: IndirectNameMap is unique by NameMapAssoc.Index, but NameMapAssoc.NameMap needn't be unique.
+// Note: When encoding in the Binary format, this must be ordered by NameMapAssoc.Index
+// https://www.w3.org/TR/wasm-core-1/#binary-indirectnamemap
+type IndirectNameMap []*NameMapAssoc
+
+type NameMapAssoc struct {
+	Index   Index
+	NameMap NameMap
 }