--- a/wasm/text/decoder.go
+++ b/wasm/text/decoder.go
@@ -27,28 +27,7 @@ func DecodeModule(source []byte) (result *wasm.Module, err error) {
 
 	// Now, handle any imported functions. Notably, we retain the same insertion order as defined in the text format in
 	// case a numeric index is used for the start function (or another reason such as the call instruction).
-	var functionNames map[uint32]string
-	var localNames map[uint32]map[uint32]string
-	for i, f := range m.importFuncs {
-		funcidx := uint32(i)
-		if f.funcName != "" {
-			if functionNames == nil {
-				functionNames = map[uint32]string{funcidx: f.funcName}
-			} else {
-				functionNames[funcidx] = f.funcName
-			}
-		}
-		if f.paramNames != nil {
-			locals := map[uint32]string{}
-			for _, pn := range f.paramNames {
-				locals[pn.index] = string(pn.name)
-			}
-			if localNames == nil {
-				localNames = map[uint32]map[uint32]string{funcidx: locals}
-			} else {
-				localNames[funcidx] = locals
-			}
-		}
+	for _, f := range m.importFuncs {
 		result.ImportSection = append(result.ImportSection, &wasm.Import{
 			Module: f.module, Name: f.name,
 			Kind:     wasm.ImportKindFunc,
@@ -56,20 +35,59 @@ func DecodeModule(source []byte) (result *wasm.Module, err error) {
 		})
 	}
 
-	// TODO: module function and local names
-
 	// The start function is called on Module.Instantiate.
 	if m.startFunction != nil {
 		result.StartSection = &m.startFunction.numeric
 	}
 
 	// Don't set the name section unless we found a name!
-	if m.name != "" || functionNames != nil || localNames != nil {
+	if localNames := mergeLocalNames(m); localNames != nil || m.name != "" || m.importFuncNames != nil {
 		result.NameSection = &wasm.NameSection{
 			ModuleName:    m.name,
-			FunctionNames: functionNames,
+			FunctionNames: m.importFuncNames,
 			LocalNames:    localNames,
 		}
 	}
 	return
+}
+
+// mergeLocalNames produces wasm.NameSection LocalNames. This has to be done post-parse as types can be defined after
+// functions that use them.
+func mergeLocalNames(m *module) wasm.IndirectNameMap {
+	j, jLen := 0, len(m.importFuncParamNames)
+	if m.typeParamNames == nil && jLen == 0 {
+		return nil
+	}
+
+	var result wasm.IndirectNameMap
+	for i, f := range m.importFuncs {
+		funcIdx := wasm.Index(i)
+
+		// seek to see if we have any function-defined parameter names
+		var inlinedNames wasm.NameMap
+		for ; j < jLen; j++ {
+			next := m.importFuncParamNames[j]
+			if next.Index > funcIdx {
+				break // we have parameter names, but starting at a later index
+			} else if next.Index == funcIdx {
+				inlinedNames = next.NameMap
+				break
+			}
+		}
+		// TODO: module function and local names
+
+		typeNames, hasType := m.typeParamNames[f.typeIndex.numeric]
+		var localNames wasm.NameMap
+		if inlinedNames == nil && !hasType {
+			continue
+		} else if inlinedNames == nil {
+			localNames = typeNames
+		} else {
+			// On conflict, choose the function names, as merge rules aren't defined in the specification. If there are
+			// names on the function, the user added them. They may not intend to inherit names they didn't define!
+			localNames = inlinedNames
+		}
+		result = append(result, &wasm.NameMapAssoc{Index: funcIdx, NameMap: localNames})
+	}
+	return result
 }