… wasm/binary/names_test.go
… wasm/example/example_test.go
… wasm/example/testdata/example.wasm
… wasm/example/testdata/example.wat
… wasm/store_test.go
… wasm/text/bind_test.go
… wasm/text/decoder_test.go
… wasm/text/parser_test.go
… wasm/text/type_parser.go
… wasm/text/types.go