This reworks the name section to be closer to the binary format

This makes the name section data structure cheaper and more closely
matches the binary structure. Doing this allows us to keep the original
decode order and avoid expense hashing or re-ordering redundantly.

This also changes language to match the word local section in
preparation of module defined functions. It also introduces `wasm.Index`
to avoid repeating docs so often, Finally, this backfills some more
tests.

As we don't re-sort etc, benchmark performance has improved, particularly in the round-trip case

Before:
```
BenchmarkCodecExample/binary.DecodeModule
BenchmarkCodecExample/binary.DecodeModule-16         	  210984	      4853 ns/op	    2784 B/op	     168 allocs/op
BenchmarkCodecExample/binary.EncodeModule
BenchmarkCodecExample/binary.EncodeModule-16         	  240454	      4977 ns/op	    3832 B/op	     143 allocs/op
BenchmarkCodecExample/text.DecodeModule
BenchmarkCodecExample/text.DecodeModule-16           	  101834	     11375 ns/op	    6600 B/op	     248 allocs/op
BenchmarkCodecExample/wat2wasm_via_text.DecodeModule->binary.EncodeModule
BenchmarkCodecExample/wat2wasm_via_text.DecodeModule->binary.EncodeModule-16         	   67006	     16439 ns/op	   10416 B/op	     391 allocs/op
BenchmarkCodecExample/wat2wasm_via_wasmtime.Wat2Wasm
BenchmarkCodecExample/wat2wasm_via_wasmtime.Wat2Wasm-16                              	   43602	     27315 ns/op	    1152 B/op	       3 allocs/op
```

Now:
```
BenchmarkCodecExample/binary.DecodeModule
BenchmarkCodecExample/binary.DecodeModule-16         	  254241	      4706 ns/op	    2304 B/op	     179 allocs/op
BenchmarkCodecExample/binary.EncodeModule
BenchmarkCodecExample/binary.EncodeModule-16         	  325231	      3603 ns/op	    3408 B/op	     127 allocs/op
BenchmarkCodecExample/text.DecodeModule
BenchmarkCodecExample/text.DecodeModule-16           	  108086	     10907 ns/op	    5928 B/op	     259 allocs/op
BenchmarkCodecExample/wat2wasm_via_text.DecodeModule->binary.EncodeModule
BenchmarkCodecExample/wat2wasm_via_text.DecodeModule->binary.EncodeModule-16         	   80340	     14828 ns/op	    9336 B/op	     386 allocs/op
```